Fix compiler warning
'argument' : conversion from 'type1' to 'type2', possible loss of data.

+++ src/client/console.h
@@ -46,7 +46,7 @@ typedef struct
 
 	int		vislines;
 
-	float	times[NUM_CON_TIMES];	// cls.realtime time the line was generated
+	int	times[NUM_CON_TIMES];	// cls.realtime time the line was generated
 								// for transparent notify lines
 } console_t;
